fix missing layout for breadcrumb
